StateMachine :Move code logic from controller to facade
(dialogue) doTransition's logic migrated to facade
